add typescript configs
