--- src/config.js
@@ -1,7 +1,0 @@
-// src/config.js
-
-export const categories = [
-  { name: 'Personal To-Do', type: 'todo', icon: 'CheckSquare', color: 'text-blue-400' },
-  { name: 'Project Idea', type: 'idea', icon: 'Lightbulb', color: 'text-yellow-400' },
-  { name: 'Meeting Note', type: 'meeting', icon: 'FileText', color: 'text-green-400' }
-];
