+++ src/config.ts
@@ -1,0 +1,7 @@
+import { Category } from './types';
+
+export const categories: Category[] = [
+  { name: 'Personal To-Do', type: 'todo', icon: 'CheckSquare', color: 'text-blue-400' },
+  { name: 'Project Idea', type: 'idea', icon: 'Lightbulb', color: 'text-yellow-400' },
+  { name: 'Meeting Note', type: 'meeting', icon: 'FileText', color: 'text-green-400' }
+];
